Move all PluginInterpolation related types into its file
